Utvid 'kan lagre og hente'-meldekort test med simulering
